refactor trace schema options
